requirements.txt
openai
python-dotenv
PyPDF2
requests
pillow
ultralytics
opencv-python
numpy
pyTelegramBotAPI
